### 2024-12-03 131f569
Add JOSS banner
