+++ paper.bib
@@ -80,7 +80,7 @@ year = {2016}
     author={Zhang, Xingguang and Chimitt, Nicholas and Chi, Yiheng and Mao, Zhiyuan and Chan, Stanley H},
     title={Spatio-Temporal Turbulence Mitigation: A Translational Perspective},
     booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
-    doi={10.1109/CVPR52733.2024.00279}
+    doi={10.1109/CVPR52733.2024.00279},
     month={June},
     year={2024}
 }
