+++ README.md
@@ -36,7 +36,7 @@ To get started with the project, follow these steps:
 - **Prerequisites:**
   In order to correctly install `speckcn2` you need `python3.9` or higher. If you don't have it installed, you can download it from the [official website](https://www.python.org/downloads/). You will also need the header files that are required to compile Python extensions and are contained in `python3-dev`. On Ubuntu, you can install them with:
   ```bash
-  apt-get install python-dev
+  apt-get install python3-dev
   ```
 
 - **Install the package:**
