### 2024-12-19 14710c5
add missing assets
