+++ src/speckcn2/plots.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import itertools
 from typing import Dict
 
 import matplotlib.pyplot as plt
@@ -129,7 +130,9 @@ def score_plot(
                  color='black')
 
     plt.tight_layout()
-    plt.savefig(f'{dirname}/single_speckle_loss{loss.item():.4g}.png')
+    figure_format = conf.get('figure_format', 'png')
+    plt.savefig(
+        f'{dirname}/single_speckle_loss{loss.item():.4g}.{figure_format}')
     plt.close()
 
 
@@ -161,7 +164,8 @@ def plot_loss(conf: dict, model, data_dir: str) -> None:
     axs.legend()
     plt.title(f'Model: {model_name}')
     plt.tight_layout()
-    plt.savefig(f'{dirname}/loss_{model_name}.png')
+    figure_format = conf.get('figure_format', 'png')
+    plt.savefig(f'{dirname}/loss_{model_name}.{figure_format}')
     plt.close()
 
 
@@ -191,7 +195,8 @@ def plot_time(conf: dict, model, data_dir: str) -> None:
     axs.legend()
     plt.title(f'Model: {model_name}')
     plt.tight_layout()
-    plt.savefig(f'{dirname}/time_{model_name}.png')
+    figure_format = conf.get('figure_format', 'png')
+    plt.savefig(f'{dirname}/time_{model_name}.{figure_format}')
     plt.close()
 
 
@@ -227,7 +232,8 @@ def plot_histo_losses(conf: dict, test_losses: list[dict],
     axs.legend()
     plt.title(f'Model: {model_name}')
     plt.tight_layout()
-    plt.savefig(f'{dirname}/histo_losses_{model_name}.png')
+    figure_format = conf.get('figure_format', 'png')
+    plt.savefig(f'{dirname}/histo_losses_{model_name}.{figure_format}')
     plt.close()
 
 
@@ -360,17 +366,18 @@ def plot_param_vs_loss(conf: dict,
                          -0.5,
                          color='tab:red',
                          alpha=0.1)
-        axs.set_xlabel(f'{name} {units}')
+        axs.set_xlabel(f'{name} {units}', fontsize=12)
         axs.set_xscale('log')
         axs.set_yscale('symlog', linthresh=0.1)
-        axs.set_ylabel('Relative error')
+        axs.set_ylabel('Relative error', fontsize=12)
         yticks = [-1, -0.5, -0.1, 0, 0.1, 0.5, 1]
         plt.yticks(yticks)
         yticklabels = ['-100%', '-50%', '-10%', '0', '10%', '50%', '100%']
         plt.gca().set_yticklabels(yticklabels)
-        plt.title(f'Model: {model_name}')
+        plt.title(f'{name} error', fontsize=15)
         plt.tight_layout()
-        plt.savefig(f'{dirname}/{param}_vs_sum_{model_name}.png')
+        figure_format = conf.get('figure_format', 'png')
+        plt.savefig(f'{dirname}/{param}_vs_sum_{model_name}.{figure_format}')
         plt.close()
 
         # If specified, plot the histogram per single bin
@@ -474,7 +481,9 @@ def plot_param_histo(conf: dict, test_losses: list[dict], data_dir: str,
         axs.legend()
         plt.title(f'Model: {model_name}')
         plt.tight_layout()
-        plt.savefig(f'{dirname}/histo_{param_true}_{model_name}.png')
+        figure_format = conf.get('figure_format', 'png')
+        plt.savefig(
+            f'{dirname}/histo_{param_true}_{model_name}.{figure_format}')
         plt.close()
 
 
@@ -556,7 +565,9 @@ def plot_J_error_details(conf: dict,
                 axs.legend()
[…]
