### 2025-02-05 d52cde9
Update figures (#67)
