### 2025-08-05 #7556 @crvernon
@observingClouds please let me know if you are able to continue as a topic editor for this submission.
