### 2025-09-30 #7556 @acsenrafilho
> 👋 [@acsenrafilho](https://github.com/acsenrafilho) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Thank you for the invitation @crvernon
I will be reviewing the material in the next few weeks
