### 2025-10-01 #9127 @editorialbot
[REVIEW]: Speckle Cn2 Profiler: Improving Satellite Communications with Machine Learning
**Submitting author:** @SCiarella (<a href="http://orcid.org/0000-0002-9247-139X">Simone Ciarella</a>)
**Repository:** https://github.com/MALES-project/SpeckleCn2Profiler.git
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.0.0
