### 2025-10-01 MALES-project/SpeckleCn2Profiler#68 @acsenrafilho
JOSS Review - Adding sample data into documentation
Hi,

I am reviewing the JOSS article: https://github.com/openjournals/joss-reviews/issues/9127

Could the authors able to add some sample data into the documentation? I think this would help new users to make a simple execution on the project locally and understand the general input/output pattern.
