### 2025-10-02 5ce3d96
Adjust figure size in paper.md
