### 2025-10-02 MALES-project/SpeckleCn2Profiler#70 @HectorMozo3110
Observation on paper (JOSS review)
👋 @SCiarella

The paper is very good overall, and I find the *Summary* clear, concise, and perfectly written for a broad audience.

However, regarding the *Statement of need*:

- At the beginning, it feels somewhat repetitive with what is already stated in the *Summary*. I would recommend analyzing and avoiding redundancies to make the narrative flow more direct.
- It would also be helpful to be more explicit about **who the target users of the software are** (e.g., aerospace engineers, astronomers, atmospheric scientists, etc.), so the community that benefits is clearly identified.
- Another important point would be to include a **more direct comparison with other related software packages**, highlighting their limitations and making clear why this tool is essential. This could also be supported with references.
- Additionally, it could be useful to add **brief references to APIs or documentation**. While the paper should not replace the documentation and should not include extensive details, adding short references would allow users to go directly from the paper to the official docs.

I also submitted a [PR #69](https://github.com/MALES-project/SpeckleCn2Profiler/pull/69) adjusting the figure (since it was causing a page break) and moving the corresponding files to the `paper` folder. If you agree, please feel free to merge it.

Finally, I have one question (if applicable): **Has this software been presented in any lab, conference, or already used in a practical research project?** If so, it would be very valuable to include a short reference to that, as it would strengthen the demonstration of its applicability.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9127)
