### 2025-10-02 MALES-project/SpeckleCn2Profiler#71 @HectorMozo3110
Observation on README (JOSS review)
👋@SCiarella

The README is already quite solid, with installation instructions, documentation links, license, usage examples, and contribution guidelines.
However, to align more fully with the JOSS [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), I suggest adding or clarifying the following:

- **Statement of need / Purpose** - while the README already explains the problem (turbulence in optical satellite communications), it would be helpful to make explicit who the target users are (e.g. aerospace engineers, atmospheric scientists, astronomers), and also to briefly mention how this tool differs from related software. This would make the “need” clearer.

- **Quick example (MWE)** - add a minimal working example directly in the README (command + expected output), so readers can quickly test functionality without leaving the page.

- **Running tests** - the documentation already explains how to run tests, but adding a short snippet in the README would make it more accessible. This way users can execute them without leaving the README, while still being able to follow the link to the full documentation for more details.

These small additions would help satisfy JOSS’s documentation expectations (especially under the “Documentation” section of the review criteria), and would also help new users get started more quickly, without having to navigate extensively.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9127)
