### 2025-10-02 MALES-project/SpeckleCn2Profiler#72 @HectorMozo3110
Suggestions for Documentation, Quickstart examples (JOSS review)
👋 @SCiarella

The documentation looks very complete and well structured.
As a small suggestion, it could include a **Quickstart section** so new users can try a minimal example right away and run quick tests.

This would make it easier to verify the installation and help users get started directly without having to go through the full documentation first.

**Note:** This observation is part of the ongoing JOSS review (see https://github.com/openjournals/jossreviews/issues/https://github.com/openjournals/joss-reviews/issues/9127)
