### 2025-10-04 MALES-project/SpeckleCn2Profiler#73 @acsenrafilho
JOSS Review - Review and Improvement Suggestions for SpeckleCn2Profiler Installation Guide
**Review**: https://github.com/openjournals/joss-reviews/issues/9127

# Commentaries for Improvement
The following changes would make the guide clearer and more user-friendly.

1. **Broaden Prerequisite Instructions**
      The current `apt-get` command is only for Debian-based systems.

      **Suggestion**: Add equivalent commands for other platforms:
      ```bash
      Fedora/RHEL: sudo dnf install python3-devel

      macOS: xcode-select --install

      Windows: Note that headers are included with the official Python installer.
      ```

2. **Mandate Virtual Environments**
      Global installations can cause dependency issues.

      **Suggestion**: Strongly recommend `venv` for all users at the start of the guide and provide a minimal code example.

3. **Add a Verification Step**
      Users need to know if the installation worked.

      **Suggestion**: Include a simple one-line command to confirm a successful installation.

      ```bash
      python -c "import speckcn2; print('Success!')"
      ```

4. **Clarify Installation Method Use Cases**
      Users may not know which installation method to choose.

      **Suggestion**: Briefly explain the target user for each method.
      ```bash
      pip: For general users.

      git clone: For developers or contributors.
      ```
