### 2025-10-04 #9127 @acsenrafilho
Hi @SCiarella

Please see the following issues created in the `SpeckleCn2Profiler` project to help in this JOSS review:
*  https://github.com/MALES-project/SpeckleCn2Profiler/issues/68
* https://github.com/MALES-project/SpeckleCn2Profiler/issues/73
