### 2025-10-06 MALES-project/SpeckleCn2Profiler#68 @SCiarella
@acsenrafilho Thank you very much!

I have added some data in the submodule containing the example scripts, and they are also referenced in the documentation for quick access.
