### 2025-10-06 6d6b8b0
Enhance install instructions
