### 2025-10-06 a22ccda
Enhance install instructions (#75)
* Enhance install instructions
