+++ docs/installation.md
@@ -1,32 +1,116 @@
 # Installation
-To get started with the project, follow these steps:
 
-- **Prerequisites:**
-  In order to correctly install `speckcn2` you need `python3.9` or higher. If you don't have it installed, you can download it from the [official website](https://www.python.org/downloads/). You will also need the Python header files that are required to compile extensions. On Ubuntu, you can install them with:
-  ```bash
-  apt-get install python3-dev
-  ```
+This guide provides step-by-step instructions for installing `speckcn2` on different platforms and setups.
 
-- **Install the package:**
-   ```bash
-   python -m pip install speckcn2
-   ```
+## Prerequisites
 
-- **Or: Clone the repository:**
-  ```bash
-  git clone https://github.com/MALES-project/SpeckleCn2Profiler.git
-  cd SpeckleCn2Profiler
-  git submodule init
-  git submodule update
-  ```
+To correctly install `speckcn2`, you need **Python 3.9 or higher**. If you don't have it installed, you can download it from the [official Python website](https://www.python.org/downloads/).
 
-## MacOS M1 arm64
+You will also need header files required to compile Python extensions (contained in `python3-dev`):
 
-Some dependencies (e.g. `scikit`) do not support the latest python version (3.12). Also `py3nj`, a dependency of `escnn`, requires openmp. We've installed this via homebrew and thus explicitly specifying the C compiler (gnu) prior to installation of this package does the trick.
+=== "Ubuntu/Debian"
+    ```bash
+    sudo apt-get install python3-dev
+    ```
 
-```sh
-conda create -n speckcn2 python=3.10
-conda activate speckcn2
-CC=gcc-13 pip3 install py3nj # install py3nj before with gcc instead of clang
+=== "Fedora/RHEL"
+    ```bash
+    sudo dnf install python3-devel
+    ```
+
+=== "macOS"
+    ```bash
+    xcode-select --install
+    ```
+
+=== "Windows"
+    Headers are included with the official Python installer - no additional steps needed.
+
+## Installation Options
+
+### Option 1: Install from PyPI (Recommended)
+
+This is the simplest way to install `speckcn2` for regular usage:
+
+```bash
+python -m pip install speckcn2
+```
+
+#### Using Virtual Environment (Strongly Recommended)
+
+We **strongly recommend** using a virtual environment to avoid dependency conflicts:
+
+=== "Linux/macOS"
+    ```bash
+    # Create virtual environment
+    python -m venv speckcn2-env
+
+    # Activate virtual environment
+    source speckcn2-env/bin/activate
+
+    # Install the package
+    python -m pip install speckcn2
+    ```
+
+=== "Windows"
+    ```cmd
+    # Create virtual environment
+    python -m venv speckcn2-env
+
+    # Activate virtual environment
+    speckcn2-env\Scripts\activate
+
+    # Install the package
+    python -m pip install speckcn2
+    ```
+
+#### Verify Installation
+
+After installation, verify that everything works correctly:
+
+```bash
+python -c "import speckcn2; print('Installation successful!')"
+```
+
+### Option 2: Development Installation
+
+For advanced users and developers who want to modify the code or contribute to the project:
+
+```bash
+git clone https://github.com/MALES-project/SpeckleCn2Profiler.git
+cd SpeckleCn2Profiler
+git submodule init
+git submodule update
 pip install -e .
 ```
+
+The `-e` flag installs the package in "editable" mode, meaning changes to the source code will be immediately reflected without reinstalling.
+
+## Platform-Specific Instructions
+
+### macOS with Apple Silicon (M1/M2)
+
+Apple Silicon Macs require special handling due to dependency compatibility issues:
+
[…]
