### 2025-10-06 MALES-project/SpeckleCn2Profiler#70 @SCiarella
@HectorMozo3110 Thank you very much for the feedback!

I’ve revised the Statement of Need to remove redundancies with the Summary, clarified the target user community, and added a more explicit comparison with related software (including references). I also included a brief reference to the documentation and API for easier access.

Regarding your question: yes, the software has been used in the STORM project, and I’ve added a short mention of this in the paper to highlight its practical use. A scientific research paper about the role of this package in the STORM project is currently under review.

I also reviewed and merged PR #69 — thank you for taking care of the figure layout issue!
