### 2025-10-06 ba5c291
Improve paper.md
