+++ paper/paper.bib
@@ -95,3 +95,60 @@ year = {2016}
   doi          = {10.5281/zenodo.14178688},
   url          = {https://doi.org/10.5281/zenodo.14178688}
 }
+
+@article{arvis2024storm,
+  title={Single Shot Line-of-Sight Atmospheric Turbulence Profiling for Laser Satellite Communications with STORM},
+  author={Arvis, Marguerite and Ciarella, Simone and Loicq, J{\'e}r{\^o}me and Saathof, Rudolf},
+  journal={Applied Optics},
+  year={2025},
+  note={In press}
+}
+
+@article{AOtools,
+author = {M. J. Townson and O. J. D. Farley and G. Orban de Xivry and J. Osborn and A. P. Reeves},
+journal = {Opt. Express},
+keywords = {Adaptive optics; Astronomical optics; Atmospheric turbulence; Deformable mirrors; Fourier transforms; Free space optics},
+number = {22},
+pages = {31316--31329},
+publisher = {Optica Publishing Group},
+title = {AOtools: a Python package for adaptive optics modelling and analysis},
+volume = {27},
+month = {Oct},
+year = {2019},
+url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-27-22-31316},
+doi = {10.1364/OE.27.031316},
+abstract = {AOtools is a Python package that is open-source and aimed at providing tools for adaptive optics users and researchers. We present version 1.0, which contains tools for adaptive optics processing, including analysing data in the pupil plane, images and point spread functions in the focal plane, wavefront sensors, modelling of atmospheric turbulence, physical optical propagation of wavefronts, and conversion between frequently used adaptive optics and astronomical units. The main drivers behind AOtools is that it should be easy to install and use. To achieve this the project features extensive documentation, automated unit testing and is registered on the Python Package Index. AOtools is under continuous active development to expand the features available, and we encourage everyone involved in adaptive optics to become involved and contribute to the project.},
+}
+
+@article{FAST,
+author = {O. J. D. Farley and M. J. Townson and J. Osborn},
+journal = {Opt. Express},
+keywords = {Adaptive optics; Atmospheric turbulence; Deformable mirrors; Free space optics; Scintillation index; Wave propagation},
+number = {13},
+pages = {23050--23064},
+publisher = {Optica Publishing Group},
+title = {FAST: Fourier domain adaptive optics simulation tool for bidirectional ground-space optical links through atmospheric turbulence},
+volume = {30},
+month = {Jun},
+year = {2022},
+url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-30-13-23050},
+doi = {10.1364/OE.458659},
+abstract = {Free space optical links between the ground and space may be severely degraded by atmospheric turbulence. Adaptive Optics, a technique allowing partial correction of this degradation, is beginning to see use in the field with the potential to achieve more robust and higher bandwidth links. Here we present a simulation tool, FAST, which utilises an analytical Fourier domain Adaptive Optics model developed for astronomy. Using the reciprocity principle, the simulation may be applied either to downlink post-compensated or uplink pre-compensated beams. We show that FAST gives similar results to full end-to-end simulations with wave-optical propagation whilst being between 10 and 200 times faster, enabling the characterisation of optical links with complex Adaptive Optics systems in timely fashion.},
+}
+
+@inproceedings{OOPAO,
+  TITLE = {{OOPAO: Object Oriented Python Adaptive Optics}},
[…]
