### 2025-10-06 MALES-project/SpeckleCn2Profiler#71 @SCiarella
@HectorMozo3110 Thanks again for the nice feedback.

I agree with your comments, so I have implemented your suggestions in the `README.md` file.
It is definitely better now, and I hope it can also satisfy JOSS' expectations.
