### 2025-10-06 MALES-project/SpeckleCn2Profiler#72 @SCiarella
Thanks @HectorMozo3110 !

I really like this suggestion! Now both the README and the docs feature a **Quickstart section**
