### 2025-10-15 #9127 @crvernon
👋  @SCiarella - we are almost there! You paper is in great shape.  I just need you to correct the following:

- [ ] LINE 141:  The "p" in "python" should be capitalized.  You can do this in your bib file by placing curly brackets around the characters you wish to maintain the formatting of.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.

After confirmation, I can then move forward with accepting the submission.
