### 2025-10-15 27fc8cf
Fix bib titles
