+++ paper/paper.bib
@@ -111,7 +111,7 @@ keywords = {Adaptive optics; Astronomical optics; Atmospheric turbulence; Deform
 number = {22},
 pages = {31316--31329},
 publisher = {Optica Publishing Group},
-title = {AOtools: a Python package for adaptive optics modelling and analysis},
+title = {{AOtools: a Python package for adaptive optics modelling and analysis}},
 volume = {27},
 month = {Oct},
 year = {2019},
