### 2025-10-15 50a26f1
JOSS Paper (#77)
* Add /paper

---------

Co-authored-by: Luisa Orozco <l.orozco@esciencecenter.nl>
Co-authored-by: Hector E Mozo <hector.mozo@neureonmindfluxlab.org>
