+++ paper/paper.bib
@@ -1,0 +1,154 @@
+@inproceedings{pytorch,
+author = {Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and Chauhan, Geeta and Chourdia, Anjali and Constable, Will and Desmaison, Alban and DeVito, Zachary and Ellison, Elias and Feng, Will and Gong, Jiong and Gschwind, Michael and Hirsh, Brian and Huang, Sherlock and Kalambarkar, Kshiteej and Kirsch, Laurent and Lazos, Michael and Lezcano, Mario and Liang, Yanbo and Liang, Jason and Lu, Yinghai and Luk, CK and Maher, Bert and Pan, Yunjie and Puhrsch, Christian and Reso, Matthias and Saroufim, Mark and Siraichi, Marcos Yukio and Suk, Helen and Suo, Michael and Tillet, Phil and Wang, Eikan and Wang, Xiaodong and Wen, William and Zhang, Shunting and Zhao, Xu and Zhou, Keren and Zou, Richard and Mathews, Ajit and Chanan, Gregory and Wu, Peng and Chintala, Soumith},
+booktitle = {29th ACM International Conference on Architectural Support for Programming Languages and Operating Systems, Volume 2 (ASPLOS '24)},
+doi = {10.1145/3620665.3640366},
+publisher = {ACM},
+title = {{PyTorch 2: Faster Machine Learning Through Dynamic Python Bytecode Transformation and Graph Compilation}},
+url = {https://pytorch.org/assets/pytorch2-2.pdf},
+year = {2024}
+}
+
+@inproceedings{cohen2016,
+  title = {Group {{Equivariant Convolutional Networks}}},
+  booktitle = {Proceedings of {{The}} 33rd {{International Conference}} on {{Machine Learning}}},
+  author = {Cohen, Taco and Welling, Max},
+  editor = {Balcan, Maria Florina and Weinberger, Kilian Q},
+  date = {2016-03},
+  volume = {48},
+  pages = {2990--2999},
+  publisher = {PMLR},
+  location = {New York, New York, USA},
+  url = {https://proceedings.mlr.press/v48/cohenc16.html},
+  abstract = {We introduce Group equivariant Convolutional Neural Networks (G-CNNs), a natural generalization of convolutional neural networks that reduces sample complexity by exploiting symmetries. G-CNNs use G-convolutions, a new type of layer that enjoys a substantially higher degree of weight sharing than regular convolution layers. G-convolutions increase the expressive capacity of the network without increasing the number of parameters. Group convolution layers are easy to use and can be implemented with negligible computational overhead for discrete groups generated by translations, reflections and rotations. G-CNNs achieve state of the art results on CIFAR10 and rotated MNIST.}
+}
+
+@online{escnn1,
+  title = {General \${{E}}(2)\$-{{Equivariant Steerable CNNs}}},
+  author = {Weiler, Maurice and Cesa, Gabriele},
+  date = {2021-04-06},
+  eprint = {1911.08251},
+  eprinttype = {arXiv},
+  doi = {10.48550/arXiv.1911.08251},
+  url = {http://arxiv.org/abs/1911.08251},
+  urldate = {2024-11-15},
[…]
