+++ paper/paper.md
@@ -1,0 +1,92 @@
+---
+title: 'Speckle Cn2 Profiler:
+Improving Satellite Communications with Machine Learning'
+tags:
+  - Python
+  - machine learning
+  - signal processing
+  - turbulence
+  - image processing
+  - equivariance
+authors:
+  - name: Simone Ciarella
+    orcid: 0000-0002-9247-139X
+    corresponding: true
+    affiliation: 1
+  - name: Luisa Orozco
+    orcid: 0000-0002-9153-650X
+    affiliation: 1
+  - name: Victor Azizi
+    orcid: 0000-0003-3535-8320
+    affiliation: 1
+  - name: Marguerite Arvis
+    orcid: 0009-0006-7409-3985
+    affiliation: 2
+  - name: Rudolf Saathof
+    orcid: 0000-0003-0368-0139
+    affiliation: 2
+
+affiliations:
+ - name: Netherlands eScience Center, Amsterdam, The Netherlands
+   index: 1
+ - name: Faculty of Aerospace Engineering, Delft University of Technology, Delft, The Netherlands
+   index: 2
+date: 13 December 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+Optical satellite communications is a growing research field in which -using lasers- signals can be sent from the ground to satellites, from satellites to satellites and then back to the ground. The main advantage of using laser communication over radio waves is increased bandwidth that enables the transfer of more data in less time.
+However, one of the challenges for this protocol is the turbulence in the atmosphere that perturbs such transmission.
+The reduction of the quality of signal communication can be calculated and then compensated, but this requires a knowledge of the turbulence strength.
+A common way to model the turbulence is to use the refractive index structure constant, $C_n^2$, which is a measure of the strength of the turbulence. Its profile can be used to estimate the effect of the turbulence on the signal and then apply a correction.
+To measure $C_n^2$, there are several alternative instruments, each one with its own limitations. One possibility is to use speckle-based observation, which consists on looking at the twinkling of the stars and use their pattern to infer the turbulence profile. This is a non-intrusive method that can be used in real time, but it requires a deep understanding of the turbulence and the observed speckle patterns, which are highly influenced by the turbulence profile.
+The connection between speckle observation, and turbulence ($C_n^2$) is not clearly understood, so an analytical theory does not exist.
+Here we present `speckcn2`, a Python package that uses machine learning to provide a numerical reconstruction of the turbulence profile from a speckle pattern [@ciarella2024].
+
+
+![Example of `speckcn2` pipeline: speckle pattern as input to output a prediction of the turbulence profile (J). \label{fig:prediction}](https://github.com/MALES-project/SpeckleCn2Profiler/blob/main/src/speckcn2/assets/single_prediction.png?raw=true){ width=90% }
+
+
+# Statement of need
+
+While deep learning has been applied to mitigate atmospheric turbulence effects in imaging through temporal mitigation using videos [@zhang2024spatio] and static image degradation compensation [@9506614], these approaches focus on visual correction rather than quantitative turbulence characterization. In contrast, `speckcn2` addresses the critical need for numerical reconstruction of $C_n^2$ profiles, enabling both scientific understanding of atmospheric turbulence and practical integration into communication system models for performance optimization.
+
+The primary target users of `speckcn2` are aerospace engineers working on optical satellite communication systems, atmospheric scientists studying turbulence phenomena, and astronomers developing adaptive optics systems. The package provides these communities with a tool to estimate turbulence profiles from speckle observations where analytical theories are lacking, using machine learning to bridge this gap.
+
[…]
