+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 # https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
 [project]
 name = "speckcn2"
-version = "1.0.1"
+version = "1.1.0"
 description = "Estimate Cn2 from Speckle patterns using machine learning"
 readme = "README.md"
 requires-python = ">=3.9"
@@ -118,7 +118,7 @@ ignore_missing_imports = true
 
 
 [tool.bumpversion]
-current_version = "1.0.1"
+current_version = "1.1.0"
 
 [[tool.bumpversion.files]]
 filename = "src/speckcn2/__init__.py"
