### 2025-10-15 66360b6
Update CITATION.cff
