+++ CITATION.cff
@@ -2,7 +2,7 @@
 # Visit https://bit.ly/cffinit to generate yours today!
 
 cff-version: 1.2.0
-title: Speckle to Cn2
+title: Speckle Cn2 Profiler: Improving Satellite Communications with Machine Learning
 message: >-
   If you use this software, please cite it using the
   metadata from this file.
@@ -13,15 +13,15 @@ authors:
     email: s.ciarella@esciencecenter.nl
     affiliation: Netherlands eScience Center
     orcid: 'https://orcid.org/0000-0002-9247-139X'
-  - given-names: Victor
-    family-names: Azizi
-    affiliation: Netherlands eScience Center
-    orcid: 'https://orcid.org/0000-0003-3535-8320'
   - given-names: Luisa
     family-names: Orozco
     email: l.orozco@esciencecenter.nl
     orcid: 'https://orcid.org/0000-0002-9153-650X'
     affiliation: Netherlands eScience Center
+  - given-names: Victor
+    family-names: Azizi
+    affiliation: Netherlands eScience Center
+    orcid: 'https://orcid.org/0000-0003-3535-8320'
   - given-names: Marguerite
     family-names: Arvis
     email: m.arvis@tudelft.nl
