+++ CITATION.cff
@@ -43,4 +43,4 @@ keywords:
   - turbulence
   - laser communication
 license: Apache-2.0
-version: 1.1.0
+version: 1.1.1
