### 2025-10-15 0b9dff0
Add JOSS paper
Update title in CITATION.cff
